+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.128.2] - 2024-12-09
+* Fix import for command `describe-service-section`
+
+## [3.128.1] - 2024-12-06
+* Fix `object of type 'NoneType' has no len()` for command `describe-service-section`
+if `'blockValue'=null`
+
 ## [3.128.0] - 2024-11-29
 * Update `SdkClient` class `execute` method. Set `UNKNOWN` instead of `SYSTEM` for signature `msg`
 * Remove `USER_IDENTIFIER = 'SYSTEM'` constant